Move errorkind package and use require package in tests
